Move frontend node graph functions into separate file
